backend/requirements.txt
django==3.2.4
djangorestframework==3.11.2
djoser==2.0.5
django_cors_headers==3.7.0
django_filters==2.4.0